Add new examples and refine models for financial calculations
This commit deletes the simple example.rs file and adds multiple new examples: volatility.rs, time-and-rate.rs, and risk-free-rate.rs. It refines the ExpirationDate handling and updates various functions in the models to improve calculations and functionality. Additionally, it introduces rayon for parallel processing to optimize performance in volatility calculations.

+++ src/model/option.rs
@@ -51,7 +51,7 @@ impl Options {
 
     pub fn time_to_expiration(&self) -> f64 {
         match self.expiration_date {
-            ExpirationDate::Days(days) => days as f64,
+            ExpirationDate::Days(days) => days,
             ExpirationDate::DateTime(datetime) => {
                 let now = Utc::now();
                 let duration = datetime.signed_duration_since(now);
